Crate concurrent_queue

A concurrent multi-producer multi-consumer queue.

There are two kinds of queues:

Bounded queue with limited capacity.

Unbounded queue with unlimited capacity.

Queues also have the capability to get closed at any point. When closed, no more items can be pushed into the queue, although the remaining items can still be popped.

These features make it easy to build channels similar to std::sync::mpsc on top of this crate.

Examples

use concurrent_queue::ConcurrentQueue;

let q = ConcurrentQueue::unbounded();
q.push(1).unwrap();
q.push(2).unwrap();

assert_eq!(q.pop(), Ok(1));
assert_eq!(q.pop(), Ok(2));

Structs

ConcurrentQueue

A concurrent queue.

Enums

PopError

Error which occurs when popping from an empty queue.

PushError

Error which occurs when pushing into a full or closed queue.